+++ tests/string.h
@@ -29,6 +29,8 @@ void unicode_na();
 
 void unicode_encoder_utf8();
 
+void unicode_encoder_utf16();
+
 } // namespace tests
 
 #endif // _TESTS_STRING_H